moved sync_timestamps to util and fixed symlink bug
